Various updates

+++ scripts/build.sh
@@ -43,21 +43,23 @@ if [ "${GOOS}x" == "windowsx" ]; then
 fi
 
 # Build needed plugins first
-ORIG_PATH=$(pwd);
-for PLUGIN_TYPE in {kms, host}; do
-    echo "==> Building ${PLUGIN_TYPE} Plugins..."
-    rm -f $ORIG_PATH/plugins/$PLUGIN_TYPE/assets/boundary-plugin-${PLUGIN_TYPE}*
-    for CURR_PLUGIN in $(ls $ORIG_PATH/plugins/$PLUGIN_TYPE/mains); do
-        cd $ORIG_PATH/plugins/$PLUGIN_TYPE/mains/$CURR_PLUGIN;
-        go build -v -o $ORIG_PATH/plugins/$PLUGIN_TYPE/assets/boundary-plugin-${PLUGIN_TYPE}-${CURR_PLUGIN}${BINARY_SUFFIX} .;
+if [ "${SKIP_PLUGIN_BUILD}x" == "x" ]; then
+    ORIG_PATH=$(pwd);
+    for PLUGIN_TYPE in {"kms","host"}; do
+        echo "==> Building ${PLUGIN_TYPE} plugins..."
+        rm -f $ORIG_PATH/plugins/$PLUGIN_TYPE/assets/boundary-plugin-${PLUGIN_TYPE}*
+        for CURR_PLUGIN in $(ls $ORIG_PATH/plugins/$PLUGIN_TYPE/mains); do
+            cd $ORIG_PATH/plugins/$PLUGIN_TYPE/mains/$CURR_PLUGIN;
+            go build -v -o $ORIG_PATH/plugins/$PLUGIN_TYPE/assets/boundary-plugin-${PLUGIN_TYPE}-${CURR_PLUGIN}${BINARY_SUFFIX} .;
+            cd $ORIG_PATH;
+        done;
+        cd $ORIG_PATH/plugins/$PLUGIN_TYPE/assets;
+        for CURR_PLUGIN in $(ls boundary-plugin*); do
+            gzip -f -9 $CURR_PLUGIN;
+        done;
         cd $ORIG_PATH;
     done;
-    cd $ORIG_PATH/plugins/$PLUGIN_TYPE/assets;
-    for CURR_PLUGIN in $(ls boundary-plugin*); do
-        gzip -f -9 $CURR_PLUGIN;
-    done;
-    cd $ORIG_PATH;
-done;
+fi
 
 if [ "${CI_BUILD}x" != "x" ]; then
     exit
